+++ src/components/Modal/Filter/index.js
@@ -1,1 +1,1 @@
-export { default } from './Filter';
+export { default } from './FilterModal';
