+++ src/components/Modal/Register/RegisterModal.jsx
@@ -5,7 +5,7 @@ import styles from '../modal.module.css';
 import Input from '#components/Input';
 import Dropdown from '#components/Dropdown';
 
-export default function Register({ option, onSubmit, onClose }) {
+export default function RegisterModal({ option, onSubmit, onClose }) {
   const [title, setTitle] = useState('');
   const [adjustmentType, setAdjustmentType] = useState(null);
   const [isOpen, setIsOpen] = useState(false);
@@ -50,7 +50,7 @@ export default function Register({ option, onSubmit, onClose }) {
   );
 }
 
-Register.propTypes = {
+RegisterModal.propTypes = {
   option: PropTypes.arrayOf(PropTypes.string).isRequired,
   onSubmit: PropTypes.func.isRequired,
   onClose: PropTypes.func.isRequired,
