+++ src/components/Modal/Register/index.js
@@ -1,1 +1,1 @@
-export { default } from './Register';
+export { default } from './RegisterModal';
