+++ src/components/Modal/Sort/SortModal.jsx
@@ -1,41 +1,36 @@
 import { useState } from 'react';
 import PropTypes from 'prop-types';
 import Modal from '../Modal';
-import styles from '../modal.module.css';
 import Dropdown from '#components/Dropdown';
+import styles from '../modal.module.css';
 
-export default function Filter({ option, onSubmit, onClose }) {
-  const [filters, setFilters] = useState([]);
+export default function SortModal({ option, onSubmit, onClose }) {
+  const [sorts, setSorts] = useState([]);
   const [selectedKey, setSelectedKey] = useState(null);
   const [selectedValue, setSelectedValue] = useState(null);
   const [isKeyOpen, setIsKeyOpen] = useState(false);
   const [isValueOpen, setIsValueOpen] = useState(false);
 
-  const handleAddFilter = () => {
+  const handleAddSort = () => {
     if (selectedKey && selectedValue) {
-      setFilters((prev) => [
-        ...prev,
-        { key: selectedKey, value: selectedValue },
-      ]);
+      setSorts((prev) => [...prev, { key: selectedKey, value: selectedValue }]);
       setSelectedKey(null);
       setSelectedValue(null);
     }
   };
 
-  const handleRemoveFilter = (index) => {
-    setFilters((prev) => prev.filter((_, i) => i !== index));
+  const handleRemoveSort = (index) => {
+    setSorts((prev) => prev.filter((_, i) => i !== index));
   };
 
-  const keys = Object.keys(option);
-
   return (
-    <Modal onSubmit={() => onSubmit?.(filters)} onClose={onClose}>
-      <span className={styles.title}>필터</span>
+    <Modal onSubmit={() => onSubmit?.(sorts)} onClose={onClose}>
+      <span className={styles.title}>정렬</span>
 
       <div className={styles.dropdownWrapper}>
         <Dropdown
-          placeHolder="필터 항목 선택"
-          options={keys}
+          placeHolder="정렬 항목"
+          options={option.keys}
           selectedValue={selectedKey}
           isOpen={isKeyOpen}
           onChange={(val) => {
@@ -49,12 +44,8 @@ export default function Filter({ option, onSubmit, onClose }) {
         />
 
         <Dropdown
-          placeHolder="값 선택"
-          options={
-            selectedKey && option[selectedKey]
-              ? option[selectedKey].options.map(String)
-              : []
-          }
+          placeHolder="정렬 방식"
+          options={option.values}
           selectedValue={selectedValue}
           isOpen={isValueOpen}
           onChange={(val) => {
@@ -69,21 +60,21 @@ export default function Filter({ option, onSubmit, onClose }) {
         <button
           type="button"
           className={styles.plusButton}
-          onClick={handleAddFilter}
+          onClick={handleAddSort}
         >
           <span className={styles.plus}>+</span>
         </button>
       </div>
 
       <div className={styles.filterWrapper}>
-        {filters.map((filter, index) => (
-          <div key={`${filter.key}-${filter.value}`} className={styles.filter}>
+        {sorts.map((sort, index) => (
+          <div key={`${sort.key}-${sort.value}`} className={styles.filter}>
             <span className={styles.filterName}>
-              {filter.key} : {filter.value}
+              {sort.key} : {sort.value}
             </span>
             <button
               type="button"
-              onClick={() => handleRemoveFilter(index)}
+              onClick={() => handleRemoveSort(index)}
               className={styles.removeButton}
             >
               <span className={styles.remove}>X</span>
@@ -95,18 +86,11 @@ export default function Filter({ option, onSubmit, onClose }) {
   );
 }
 
-Filter.propTypes = {
-  option: PropTypes.objectOf(
-    PropTypes.shape({
-      options: PropTypes.arrayOf(
-        PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-      ).isRequired,
-      currentSelectedValue: PropTypes.oneOfType([
-        PropTypes.string,
-        PropTypes.number,
-      ]),
-    }),
-  ).isRequired,
+SortModal.propTypes = {
+  option: PropTypes.shape({
+    keys: PropTypes.arrayOf(PropTypes.string).isRequired,
+    values: PropTypes.arrayOf(PropTypes.string).isRequired,
+  }).isRequired,
   onSubmit: PropTypes.func.isRequired,
   onClose: PropTypes.func.isRequired,
 };
